Update Repo Using OSS Best Practices and Clean Up
