+++ .github/ISSUE_TEMPLATE/bug-report.md
@@ -1,0 +1,33 @@
+---
+name: Bug Report
+about: File a bug for something not working as expected.
+title: ''
+labels: kind/bug, triage/needs-triage
+assignees: ''
+
+---
+<!-- Before creating a new issue please search for any existing related [issues](https://github.com/symblai/getting-started-samples/issues) before creating a new one. If you find an issue, join the discussion or add an emoji or comment with your scenario or subscribe to follow updates. -->
+
+## Bug Report
+
+<!-- Please concisely describe the problem you are having. -->
+
+## Expected Behavior
+
+<!-- Describe what you expected to happen. -->
+
+## Steps to Reproduce the Bug
+
+<!-- Describe steps to reproduce here and how reproducible this issue is (occasionally, often, always) -->
+
+## Screenshots or additional information and context
+
+<!-- If screenshots or snippets can be provided, please do so here. Otherwise please add additional context if necessary. -->
+
+## Environment Details
+
+* Operating System:
+* Node.js version (`node --version`): 
+
+## Diagnostics logs
+<!-- Collecting diagnostics logs that can help us troubleshoot your issue -->
