+++ .github/ISSUE_TEMPLATE/documentation.md
@@ -1,0 +1,20 @@
+---
+name: Docs Request
+about: File a request for incorrect or suggested docs.
+title: ''
+labels: kind/docs, triage/needs-triage
+assignees: ''
+
+---
+
+## Description
+
+<!-- Please add a detailed description of the issue or suggestion -->
+
+## Location
+
+<!-- If this is related to an existing page, please include a link to the current page so we can make sure we know where to look -->
+
+## Additional Context
+
+<!-- Any additional context that would be useful -->
