+++ .github/ISSUE_TEMPLATE/feature-request.md
@@ -1,0 +1,23 @@
+---
+name: Feature Request
+about: File a request for a new feature.
+title: ''
+labels: kind/feature, triage/needs-triage
+assignees: ''
+
+---
+
+<!-- Before creating a new feature request please search for any existing related [issues](https://github.com/symblai/getting-started-samples/issues). If you find an issue, join the discussion or add an emoji or comment with your scenario or subscribe to follow updates. -->
+
+## Feature Request
+
+<!-- Provide details on the feature request here. Is your feature request relating to fixing a bug or is it to add/remove capabilities? Please describe what you would like to happen. -->
+
+## Describe alternatives you've considered
+
+<!-- if applicable, what other alternatives are there and why did you not choose them? -->
+
+## Additional context
+
+<!-- if applicable, put any additional context here. -->
+
