+++ .gitignore
@@ -2,3 +2,4 @@
 *.iml
 node_modules
 .env
+.DS_Store
