+++ README.md
@@ -2,7 +2,7 @@
 
 Symbl's APIs empower developers to enable: 
 - **Real-time** analysis of free-flowing discussions to automatically surface highly relevant summary discussion topics, contextual insights, suggestive action items, follow-ups, decisions, and questions.
-- **Voice APIs** that makes it easy to add AI-powered conversation intelligence to either [telephony][telephony] or [WebSocket][websocket] interfaces.
+- **Voice APIs** that makes it easy to add AI-powered conversation intelligence to either [Telephony][telephony], [Streaming][streaming], or [Async][async] interfaces.
 - **Conversation APIs** that provide a REST interface for managing and processing your conversation data.
 - **Summary UI** with a fully customizable and editable reference experience that indexes a searchable transcript and shows generated actionable insights, topics, timecodes, and speaker information.
 
@@ -16,7 +16,7 @@ Symbl's APIs empower developers to enable:
     can see how to initialize the Voice SDK, connect to the endpoint, push
     speaker events and get the summary URL.
     [Read More](./examples/voice-sdk/telephony-speaker-events/Readme.md)
-- [Realtime WebSocket using Voice SDK](./examples/voice-sdk/realtime-websocket-single-stream/index.js)
+- [Realtime WebSocket using Voice SDK](./examples/voice-sdk/realtime-websocket-topics/index.js)
   - Connect directly using WebSocket API and push audio and get results in
     real-time. Including Live Transcript and Insights.
     [Read More](./examples/voice-sdk/realtime-websocket-single-stream/Readme.md)
@@ -114,13 +114,14 @@ If you have any questions, feel free to reach out to us at devrelations@symbl.ai
 
 This guide is actively developed, and we love to hear from you! Please feel free to [create an issue][issues] or [open a pull request][pulls] with your questions, comments, suggestions and feedback.  If you liked our integration guide, please star our repo!
 
-This library is released under the [Apache License][license]
+This library is released under the [MIT License][license]
 
-[license]: LICENSE.txt
 [telephony]: https://docs.symbl.ai/docs/telephony/overview/post-api
-[websocket]: https://docs.symbl.ai/docs/streamingapi/overview/introduction
+[streaming]: https://docs.symbl.ai/reference/streaming-api-reference
+[async]: https://docs.symbl.ai/reference/async-api-1
 [developer_community]: https://community.symbl.ai/?_ga=2.134156042.526040298.1609788827-1505817196.1609788827
 [slack]: https://join.slack.com/t/symbldotai/shared_invite/zt-4sic2s11-D3x496pll8UHSJ89cm78CA
 [signup]: https://platform.symbl.ai/?_ga=2.63499307.526040298.1609788827-1505817196.1609788827
 [issues]: https://github.com/symblai/getting-started-samples/issues
 [pulls]: https://github.com/symblai/getting-started-samples/pulls
+[license]: LICENSE
