--- examples/voice-sdk/realtime-websocket-single-stream/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "getting-started-voice-sdk",
-  "version": "1.0.0",
-  "description": "Quick code samples to get started and explore Symbl's APIs",
-  "main": "index.js",
-  "directories": {
-    "example": "examples"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/symblai/getting-started-voice-sdk.git"
-  },
-  "author": "",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/symblai/getting-started-samples/issues"
-  },
-  "homepage": "https://github.com/symblai/getting-started-samples#readme",
-  "dependencies": {
-    "dotenv": "^8.2.0",
-    "mic": "^2.1.2",
-    "request": "^2.88.2",
-    "symbl-node": "^1.0.13",
-    "uuid": "^8.1.0"
-  }
-}
