+++ package.json
@@ -17,10 +17,11 @@
   },
   "homepage": "https://github.com/symblai/getting-started-samples#readme",
   "dependencies": {
+    "@symblai/symbl-js": "^1.4.8",
     "dotenv": "^8.2.0",
     "mic": "^2.1.2",
+    "request": "^2.88.2",
     "symbl-node": "^1.0.11",
-    "uuid": "^8.1.0",
-    "request": "^2.88.2"
+    "uuid": "^8.1.0"
   }
 }
